task_manager_django/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==1.7.1
sqlparse==0.5.1
tzdata==2024.1
